requirements.txt
Flask==2.3.2
requests==2.31.0
SQLAlchemy==2.0.19
Flask-SQLAlchemy==3.0.5
psycopg2-binary==2.9.7
python-dotenv==1.0.0
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
flask-cors==4.0.0
flask-migrate==4.0.4
gunicorn==21.2.0
celery==5.3.1
redis==4.6.0 